generalize scala_proto_srcjar so that it takes flags and generator
